Stable version

+++ abstractmetadata.h
@@ -11,7 +11,7 @@ public:
 
     virtual QDateTime pictureDate () = 0;
     virtual int rating () = 0;
-    virtual int orientation () = 0;
+    virtual quint16 orientation () = 0;
 
     virtual bool hasGpsInfo () = 0;
     virtual double gpsLatitude () = 0;
@@ -21,7 +21,7 @@ public:
     virtual double gpsAltitude () = 0;
 
     virtual void setRating (int value) = 0;
-    virtual void setOrientation (int value) = 0;
+    virtual void setOrientation (quint16 value) = 0;
 };
 
 #endif // ABSTRACTMETADATA_H
