+++ xmpmetadata.h
@@ -14,7 +14,7 @@ public:
 
     QDateTime pictureDate ();
     int rating ();
-    int orientation ();
+    quint16 orientation();
 
     bool hasGpsInfo();
     double gpsLatitude ();
@@ -24,7 +24,7 @@ public:
     double gpsAltitude ();
 
     void setRating (int value);
-    void setOrientation (int value);
+    void setOrientation (quint16 value);
 
 private:
     Exiv2::Image::AutoPtr _fileData;
